Use .buckconfig.local instead of .buckconfig
